Componentizes reducers.
