+++ src/redux/reducers/genres.reducer.js
@@ -1,0 +1,9 @@
+const genres = (state = [], action) => {
+  switch (action.type) {
+    case 'SET_GENRES':
+      return action.payload;
+    default:
+      return state;
+  }
+};
+export default genres;
