+++ src/redux/reducers/singleMovie.reducer.js
@@ -1,0 +1,10 @@
+const singleMovie = (state = {}, action) => {
+  switch (action.type) {
+    case 'SET_SINGLE_MOVIE':
+      return action.payload;
+    default:
+      return state;
+  }
+};
+
+export default singleMovie;
